fix (front-end and 2 backend canisters as well) : correct code in all canister for (proposals , follow dao, unfollow, dao ), inster canister calls
