+++ src/dao_canister/src/functions.rs
@@ -2,7 +2,6 @@ mod proposal_functions;
 use candid::{CandidType, Principal};
 use ic_cdk::api::call::{CallResult, RejectionCode};
 mod dao_functions;
-// pub use dao_functions::*;
 
 mod ledger_functions;
 pub use ledger_functions::*;
