+++ src/daohouse_frontend/src/pages/FeedPage/ProposalDetails.jsx
@@ -138,7 +138,6 @@ const ProposalsDetails = () => {
   
       const joinDaoPayload = {
         place_to_join: place_to_join,
-        daohouse_backend_id: daohouseBackendId,
       };
       
       const response = await daoActor.ask_to_join_dao(joinDaoPayload);
@@ -146,7 +145,7 @@ const ProposalsDetails = () => {
       const sound = new Audio(messagesound)
       if (response.Ok) {
         setJoinStatus("Requested");
-        toast.success(res.Ok);
+        toast.success(response.Ok);
         sound.play();
       } else {
         console.error( response.Err );
@@ -173,8 +172,8 @@ const ProposalsDetails = () => {
     try {
       const daoActor = createDaoActor(daoCanisterId);
       const response = isFollowing
-        ? await daoActor.unfollow_dao(backendCanisterId)
-        : await daoActor.follow_dao(backendCanisterId);
+        ? await daoActor.unfollow_dao()
+        : await daoActor.follow_dao();
   
         if (response?.Ok) {
           toast.success(newIsFollowing ? "Successfully followed" : "Successfully unfollowed");
